move other sights models to sight folder
