+++ lib/data/model/sights/searched_sight.dart
