+++ lib/data/model/sights/sight_filter.dart
