+++ lib/ui/screen/sight_search_screen/sight_search_screen.dart
@@ -3,9 +3,9 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:places/constants/app_strings.dart';
 import 'package:places/data/model/enums/search_screen_state_type.dart';
-import 'package:places/data/model/searched_sight.dart';
+import 'package:places/data/model/sights/searched_sight.dart';
 import 'package:places/data/model/sights/sight.dart';
-import 'package:places/data/model/sight_filter.dart';
+import 'package:places/data/model/sights/sight_filter.dart';
 import 'package:places/mocks.dart';
 import 'package:places/styles/custom_icons.dart';
 import 'package:places/ui/screen/sight_search_screen/search_hystory_tile.dart';
